+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Current
 =======
 
+- Rename :code:`callable` parameters to :code:`func`.
 - Remove :code:`TrialParameters.apply` to :code:`TrialParameters.call`.
 - Remove :code:`Experiment.post_grid`.
 - Make scikit-learn and pandas optional dependencies.